Add Assertions
There are several instances where the Storage may read a task from the
file as `null`. This adds assertions so that those scenarios will not
fail silently.

As the input of data into the file is strictly sanitised, such failures
will only occur if the user's computer errors.
