--- src/main/java/rick/Launcher.java
@@ -1,20 +1,0 @@
-package rick;
-
-import javafx.application.Application;
-
-/**
- * The main class to run the rick.Rick App.
- *
- * @author SeeuSim
- *         AY2223-S2 CS2103T
- */
-public class Launcher {
-    /**
-     * The main method of the Rick App.
-     * Command line configurations may be passed in here.
-     * @param args The command line options for this app.
-     */
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
