+++ src/main/java/rick/Main.java
@@ -1,44 +1,20 @@
 package rick;
 
-
-import java.io.IOException;
-
 import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-import rick.gui.MainWindow;
 
 /**
- * A GUI for Rick using FXML.
+ * The main class to run the rick.Rick App.
  *
  * @author SeeuSim
  *         AY2223-S2 CS2103T
  */
-public class Main extends Application {
-
-    private final Rick rick = new Rick();
-
+public class Main {
     /**
-     * The default method to initialize this App's GUI.
-     *
-     * @param stage The primary stage for this application, onto which
-     *              the application scene can be set. Applications may
-     *              create other stages, if needed, but they will not be
-     *              primary stages.
+     * The main method of the Rick App.
+     * Command line configurations may be passed in here.
+     * @param args The command line options for this app.
      */
-    @Override
-    public void start(Stage stage) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setRick(rick);
-            stage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+    public static void main(String[] args) {
+        Application.launch(MainApp.class, args);
     }
 }
